+++ src/serious_python/example/flet_example/android/app/src/main/kotlin/com/example/flet_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.flet_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()